+++ webui/react/src/pages/WorkspaceDetails/ProjectCard.tsx
@@ -1,7 +1,7 @@
 import { Tooltip, Typography } from 'antd';
 import React, { useCallback } from 'react';
 
-import Avatar from 'components/Avatar';
+import Avatar from 'shared/components/Avatar/Avatar';
 import Link from 'components/Link';
 import TimeAgo from 'components/TimeAgo';
 import { paths } from 'routes/utils';
