+++ webui/react/src/shared/components/Avatar/Avatar.module.scss
