+++ webui/react/src/shared/components/Avatar/Avatar.stories.tsx
