+++ webui/react/src/shared/components/Avatar/Avatar.test.tsx
